+++ rustworkx-core/src/dag_algo.rs
